+++ src/query/QueryContinuId.js
@@ -60,8 +60,8 @@ export default class QueryContinuId extends Query {
   constructor (graphQLClient, knishIOClient) {
     super(graphQLClient, knishIOClient)
 
-    this.$__query = gql`query ($bundle: String!) {
-      ContinuId(bundle: $bundle) {
+    this.$__query = gql`query ($bundle: String!, $token: String) {
+      ContinuId(bundle: $bundle, token: $token) {
         address,
         bundleHash,
         tokenSlug,